chatbot.py
# chatbot.py

# Import the necessary libraries
import nltk
import pytesseract
from nltk.chat.util import Chat, reflections
from django.http import JsonResponse
from django.views.decorators.csrf import csrf_exempt
import json
from PyPDF2 import PdfFileReader
import docx
import cv2

# Define responses for the chatbot
pairs = [
    # Standard greetings
    ["Hola", ["¡Hola!", "¿En qué puedo ayudarte hoy?"]],
    ["Hola (.*)", ["Hola %1, ¿en qué puedo ayudarte hoy?"]],
    ["¿Cómo estás?", ["Estoy bien, gracias. ¿En qué puedo ayudarte?"]],
    
    # DuckDuckGo integration (add your API calls here)
    ["Buscar en DuckDuckGo (.*)", ["Realizando búsqueda en DuckDuckGo para %1..."]],
    
    # PDF processing
    ["Procesar PDF (.*)", ["Procesando PDF %1..."]],
    
    # Word processing
    ["Procesar Word (.*)", ["Procesando documento Word %1..."]],
    
    # Excel processing
    ["Procesar Excel (.*)", ["Procesando archivo Excel %1..."]],
    
    # PowerPoint processing
    ["Procesar PowerPoint (.*)", ["Procesando presentación PowerPoint %1..."]],
    
    # Text recognition using OpenCV
    ["Reconocer texto en imagen", ["Cargando imagen para reconocimiento de texto..."]],
    
    # Learning capabilities (implement your learning logic here)
    ["Aprender", ["Aprendiendo nuevas habilidades automáticamente..."]],
    
    # Default response
    ["(.*)", ["Lo siento, no entiendo. ¿Puedes ser más específico?"]],
]

# Create a chatbot instance
chatbot = Chat(pairs, reflections)

@csrf_exempt
def chatbot_view(request):
    # Get the user's message from the POST request
    user_message = json.loads(request.body.decode('utf-8'))['message']
    
    # Get the chatbot's response
    bot_response = chatbot.respond(user_message)
    
    # Return the response as JSON
    return JsonResponse({'message': bot_response})

# Define routes and API endpoints for the web interface (Django or Flask)
# ...

# Main program (for testing)
if __name__ == "__main__":
    nltk.download("punkt")  # Download NLTK data
    pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'  # Set Tesseract path
    # Start the Django or Flask web server here (if using a web interface)
